+++ react/__mocks__/vtex.pixel-manager/PixelContext.js
@@ -1,0 +1,6 @@
+export function usePixel(Comp) {
+  const usePixel = () => {
+    const push = () => {}
+  }
+  return usePixel
+}